rdatagouv: Tools to download and explore datasets from the French open data platform

This package provides a small client for the public API of data.gouv.fr, the French government's open data platform. It lets you list the published datasets, download their tabular resources and compute summary metrics (file weight, number of columns, missing-value rates, ...). Requests are built on top of the httr2 package.

See also

Author

Maintainer: Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Authors:

Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Magali Berland magali.berland@inrae.fr

Pierre Gloaguen pierre.gloaguen@univ-ubs.fr

Arthur Leroy arthur.leroy@inrae.fr

Mahendra Mariadassou mahendra.mariadassou@inrae.fr

Cédric Midoux cedric.midoux@inrae.fr

Jean-François Rey jean-francois.rey@inrae.fr